requirements.txt
Flask
yt-dlp
python-dotenv
gunicorn
mutagen
celery[redis]
redis
pydantic
